fix: wrong type of CartItem price
- price(Int) -> price(Double)

+++ ios/Classes/models/CartItem.swift
@@ -7,9 +7,9 @@
 
 struct CartItem {
   var name: String
-  var price: Int
+  var price: Double
   
-  init(name: String, price: Int) {
+  init(name: String, price: Double) {
     self.name = name
     self.price = price
   }
@@ -23,7 +23,7 @@ struct CartItem {
   
   static func fromDictionary(_ dictionary: [String: Any]) -> CartItem? {
     guard let name = dictionary["name"] as? String,
-          let price = dictionary["price"] as? Int else {
+          let price = dictionary["price"] as? Double else {
       return nil
     }
